Handle empty configuration parameters in integrations
When users unset integrations in the main app, the setting strings
currently become blank instead of the configuration becoming disabled,
and since all configuration is handled by the integrations gem it's hard
to distinguish between un-setting and changing an integration
setting. This makes the integration gem robust to empty settings,
although it's maybe not the cleanest solution.